Extract test_job_api.py from test_project_api.py
